% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{add_random_confounders}
\alias{add_random_confounders}
\title{Add confounders to a DAG}
\usage{
add_random_confounders(dag, prob_confound)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}

\item{prob_confound}{Numeric --- between 0 and 1. The probability that
a pair of observed nodes is confounded.}
}
\value{
List. The list is made of:
\itemize{
\item \code{dag_confounders} --- Square binary matrix. Represents the full
DAG made of observed and hidden variables.
\item \code{pos_confounders} --- Integer vector. Represents the position
of confounders (rows and columns) in \code{dag_confounders}.
}
}
\description{
Add confounders (i.e., hidden variables) at random to the DAG \code{dag}.
Each confounder is a source node and acts only on two observed variables.
The probability that a pair of observed variables is confounded is set by
\code{prob_confound}.
}
