% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{broken_hockeystick}
\alias{broken_hockeystick}
\title{Apply shifted-hockeystick function}
\usage{
broken_hockeystick(v, q_low = 0.05, q_high = 0.95)
}
\arguments{
\item{v}{Numeric vector.}

\item{q_low}{Numeric, between 0 and 1.}

\item{q_high}{Numeric, between 0 and 1.}
}
\value{
Numeric vector.
}
\description{
Apply shifted-hockeystick function to a numeric vector \code{v}.
The shifted-hockeystick function sets to zero all entries of
\code{v} greater or equal to the \code{q_low}-quantile and
strictly less than the \code{q_high}-quantile.
}
