% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{caus_order_to_adjmat}
\alias{caus_order_to_adjmat}
\title{Convert a causal order into a fully connected DAG}
\usage{
caus_order_to_adjmat(caus_order)
}
\arguments{
\item{caus_order}{Numeric vector.
Represents a causal order.}
}
\value{
Numeric matrix. The adjacency matrix of the fully connected DAG.
}
\description{
Convert the given causal order \code{caus_order} into a fully connected
DAG.
}
