% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{caus_order_to_cpdag}
\alias{caus_order_to_cpdag}
\title{Convert a causal order into a CPDAG}
\usage{
caus_order_to_cpdag(caus_order)
}
\arguments{
\item{caus_order}{Numeric vector.
Represents a causal order.}
}
\value{
Square binary matrix. A CPDAG that agrees
with the given DAG.
}
\description{
Convert the given causal order \code{caus_order} into a CPDAG
DAG.
}
