% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_functions.R
\name{causal_discovery}
\alias{causal_discovery}
\title{Perform causal discovery}
\usage{
causal_discovery(method = c("fast", "full", "greedy", "lingam", "maxmin",
  "minimax", "oracle", "pc", "random")[6], ...)
}
\arguments{
\item{method}{String. Is one of:
\itemize{
\item \code{"fast"}, see \code{\link{fast_perm_search}}.
\item \code{"full"}, see \code{\link{full_perm_search}}.
\item \code{"greedy"}, see \code{\link{greedy_perm_search}}.
\item \code{"lingam"}, see \code{\link{lingam_search}}.
\item \code{"maxmin"}, see \code{\link{fast_perm_search}}.
\item \code{"minimax"}, see \code{\link{minimax_search}}.
\item \code{"oracle"}, see \code{\link{oracle_search}}.
\item \code{"pc"}, see \code{\link{pc_search}}.
\item \code{"random"}, see \code{\link{random_perm_search}}.
}}

\item{...}{The argument for the given \code{method}. Provide:
\itemize{
\item \code{delta = <delta_matrix>},
 when \code{method} is \code{"fast", "full", "greedy", "maxmin"}.
\item \code{gamma = <gamma_matrix>},
 when \code{method} is \code{"minimax"}.
\item \code{mat = <observations_matrix>},
 when \code{method} is \code{"lingam", "pc"}.
\item \code{g = <adjacency_matrix>},
 when \code{method} is \code{"oracle", "random"}.
}}
}
\value{
List. The list is made of:
\itemize{
\item \code{est_g} --- Numeric matrix (or \code{NA} in case of error).
The estimated adjacency matrix of a DAG or CPDAG.
\item \code{order} --- Numeric vector (or \code{NA} in case of error).
The causal order of the DAG \code{est_g}. If \code{est_g} is a CPDAG,
it is first converted to a DAG in order to obtain the causal order.
}
}
\description{
Perform causal discovery using a particular \code{method}.
}
