% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{causal_tail_coeff}
\alias{causal_tail_coeff}
\title{Estimate causal tail coefficient}
\usage{
causal_tail_coeff(v1, v2, k = floor(2 * n^0.4), to_rank = TRUE,
  both_tails = TRUE)
}
\arguments{
\item{v1, v2}{Numeric vectors. Two vectors with \code{n} observations.}

\item{k}{Integer. The number of extreme observations used to
compute the causal tail coefficient. Set by default to
\code{k = floor(2 * n^0.4)}.}

\item{to_rank}{Boolean. Are the vectors \code{v1} and \code{v2}
already sorted?
Set by default to \code{TRUE}.}

\item{both_tails}{Boolean. Do you consider both tails when computing the
causal tail coefficient? If \code{both_tails = TRUE}, then you use the
\eqn{\Psi} formulation. If \code{both_tails = FALSE},
then you use the \eqn{\Gamma} formulation.
Set by default to \code{TRUE}.}
}
\value{
Numeric --- between 0 and 1.
The causal tail coefficient between \code{v1} and \code{v2}.
}
\description{
Estimates the causal tail coefficient between two variables
\code{v1} and \code{v2}, given the threshold \code{k}.
}
\details{
The causal tail coefficient is defined in
the paper "Causality in heavy-tailed models" and has two formulations.
\enumerate{
\item The first formulation is defined in the paper as the
\eqn{\Gamma}-coefficient, and it considers only the upper tails of the
variables. To use this formulation set \code{both_tails = FALSE}.
\item The second formulation is defined in the paper as the
\eqn{\Psi}-coefficient, and considers both tails of the variables.
To use this formulation set \code{both_tails = FALSE}.
}
In general, the causal tail coefficient is not symmetric, i.e.,
\code{causal_tail_coeff(v1, v2) != causal_tail_coeff(v2, v1)}.
}
