% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{causal_tail_matrix}
\alias{causal_tail_matrix}
\title{Estimate causal tail matrix}
\usage{
causal_tail_matrix(dat, k = floor(2 * n^0.4), both_tails = TRUE)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{k}{Integer. The number of extreme observations used to
compute the causal tail coefficient. Set by default to
\code{k = floor(2 * n^0.4)}.}

\item{both_tails}{Boolean. Do you consider both tails when computing the
causal tail coefficient? If \code{both_tails = TRUE}, then you use the
\eqn{\Psi} formulation. If \code{both_tails = FALSE},
then you use the \eqn{\Gamma} formulation.
Set by default to \code{TRUE}.}
}
\value{
Numeric matrix. The entry \eqn{(i, j)} contains the
causal tail coefficient between the \eqn{i}-th and the \eqn{j}-th
column of \code{dat}. The values on the main diagonal are set to \code{NA}.
}
\description{
Estimates the matrix with causal tail coefficients,
given a dataset \code{dat}, and a threshold \code{k}.
}
\details{
For more information see the documentation of
\code{\link{causal_tail_coeff}}.
}
