% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{check_caus_order}
\alias{check_caus_order}
\title{Check the causal order against a DAG}
\usage{
check_caus_order(caus_order, g)
}
\arguments{
\item{caus_order}{Numeric vector (or vector of \code{NA}s).
Represents a causal order.}

\item{g}{Numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Logical (or \code{NA}, if \code{caus_order} is a vector
of \code{NA}s).
}
\description{
Produces \code{TRUE} if \code{caus_order} is a causal order of the DAG
\code{g}, \code{FALSE} otherwise.
}
