% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{compute_ancestral_distance}
\alias{compute_ancestral_distance}
\title{Compute ancestral distance}
\usage{
compute_ancestral_distance(dag, caus_order)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}

\item{caus_order}{Numeric vector.
Represents a causal order.}
}
\value{
Numeric --- between 0 and 1. The ancestral distance between the
DAG \code{dag} and the causal order \code{caus_order}.
}
\description{
Computes the ancestral distance between the DAG \code{dag} and
the causal order \code{caus_order}. The ancestral distance is defined
as the number of pairs \eqn{(i, j)} where \eqn{i} is an ancestor
of \eqn{j} and \eqn{i} is placed after \eqn{j} in the causal order
\code{caus_order}.
}
