% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{compute_gamma}
\alias{compute_gamma}
\title{Causal gamma coefficient}
\usage{
compute_gamma(v1, v2, k = floor(2 * n^0.4), to_rank = TRUE,
  both_tails = TRUE)
}
\arguments{
\item{v1, v2}{Numeric vectors. Two vectors with \code{n} observations.}

\item{k}{Integer. The number of extreme observations used to
compute the gamma coefficient. Set by default to \code{k = floor(2 * n^0.4)}.}

\item{to_rank}{Boolean. Are the vectors \code{v1} and \code{v2} already sorted?
Set by default to \code{TRUE}.}

\item{both_tails}{Boolean. Do you consider both tails when computing gamma?
Set by default to \code{TRUE}.}
}
\value{
Numeric --- between 0.5 and 1. The gamma causal coefficient between \code{v1} and
\code{v2}.
}
\description{
Computes the causal gamma coefficient between \code{v1} and \code{v2},
given the threshold \code{k}. In general, the gamma coefficient is not
symmetric, i.e., \code{compute_gamma(v1, v2) != compute_gamma(v2, v1)}.
}
