% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{compute_gamma_matrix}
\alias{compute_gamma_matrix}
\title{Causal gamma matrix}
\usage{
compute_gamma_matrix(mat, k = floor(sqrt(n)))
}
\arguments{
\item{mat}{Numeric matrix. Observation matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{k}{Integer. The number of upper-order statistics used to
compute the gamma coefficient. Set by default to \code{k = floor(sqrt(n))}.}
}
\value{
Numeric matrix. The entry \eqn{(i, j)} contains the
gamma coefficient between the \eqn{i}-th and the \eqn{j}-th
column of \code{mat}. If \eqn{i = j}, the value of the entry \eqn{(i, j)}
is set to \code{NA}.
}
\description{
Computes the gamma coefficient matrix given a dataset \code{mat},
and a threshold \code{k}.
}
