% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{compute_gamma_theo}
\alias{compute_gamma_theo}
\title{Pairwise theoretical gamma coefficient}
\usage{
compute_gamma_theo(wg, i, j, alpha, noise_w)
}
\arguments{
\item{wg}{Numeric matrix. The weighted adjacency matrix of a DAG.
All the entries must be non-negative.}

\item{i, j}{Integer. The indices of the nodes to consider to compute
the theoretical gamma coefficient.}

\item{alpha}{Numeric. The tail-index of the noise distribution}

\item{noise_w}{Numeric vector. A \eqn{p}-dimensional vector containing
the scaling coefficients of the noise variables. Note that \eqn{p} is the
number of variables (nodes) in the DAG. By default, all entries are
set equal to one.}
}
\value{
Numeric --- between 0.5 and 1. The theoretical gamma coefficient
between node \code{i} and \code{j} in the DAG \code{wg}.
}
\description{
Computes the theoretical gamma coefficient between node \code{i}
and node \code{j}, given the weighted adjacency matrix \code{wg},
the tail-index \code{alpha} and the scales of the
noise variables \code{noise_w}.
}
