% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{compute_gamma_theo_matrix}
\alias{compute_gamma_theo_matrix}
\title{Theoretical gamma coefficient of a DAG}
\usage{
compute_gamma_theo_matrix(g, alpha, noise_w)
}
\arguments{
\item{g}{Numeric matrix. The weighted adjacency matrix of a DAG.
All the entries must be non-negative.}

\item{alpha}{Numeric. The tail-index of the noise distribution}

\item{noise_w}{Numeric vector. A \eqn{p}-dimensional vector containing
the scaling coefficients of the noise variables. Note that \eqn{p} is the
number of variables (nodes) in the DAG. By default, all entries are
set equal to one.}
}
\value{
Numeric matrix. The entry \eqn{(i, j)} contains the theoretical
gamma coefficient between the \eqn{i}-th and the \eqn{j}-th node
of \code{g}. If \eqn{i = j}, the value of the entry \eqn{(i, j)} is set
to \code{NA}.
}
\description{
Computes the theoretical gamma coefficient matrix given the DAG \code{g},
the tail-index \code{alpha} and the scales of the noise variables
\code{noise_w}.
}
