% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{compute_str_ham_distance}
\alias{compute_str_ham_distance}
\title{Compute structural Hamming distance}
\usage{
compute_str_ham_distance(cpdag, est_cpdag)
}
\value{
Numeric --- between 0 and 1. The structural Hamming
distance between a true CPDAG \code{cpdag} and the
estimated CPDAG \code{est_cpdag}.
}
\description{
Compute the structural Hamming distance between the true CPDAG
\code{cpdag} and the estimated CPDAG \code{est_cpdag}.
The structural Hamming distance is defined as in the paper from
Tsamardinos I., Brown L.E., and Aliferis C.F.,
\url{https://link.springer.com/article/10.1007/s10994-006-6889-7}.
}
