% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{dag_to_cpdag}
\alias{dag_to_cpdag}
\title{Convert a DAG into a CPDAG}
\usage{
dag_to_cpdag(dag)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}
}
\value{
Square binary matrix. A CPDAG that agrees
with the given DAG.
}
\description{
Convert the given DAG \code{dag} into a CPDAG
DAG.
}
