% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{direct_lingam_search}
\alias{direct_lingam_search}
\title{DirectLiNGAM search}
\usage{
direct_lingam_search(dat)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}
}
\value{
Numeric vector (or \code{NA} in case of error).
The causal order estimated from the data.
}
\description{
Runs DirectLiNGAM given a dataset \code{dat}. Returns the causal order
estimated from the data using the Direct LiNGAM algorithm.
}
\details{
This function performs the performs pairwise~LiNGAM algorithm of Hyvarinen
and Smith (2013, JMLR) in the DirectLiNGAM framework of Shimizu et al. (2011, JMLR).
The function calls internal C++ code.
}
