% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{ease}
\alias{ease}
\title{Extremal Ancestral SEarch}
\usage{
ease(dat, k = floor(n^0.4), both_tails = TRUE)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{k}{Positive integer. The number of extreme observations used to
compute the causal tail coefficient. Set by default to
\code{k = floor(n^0.4)}. It must be greater than 1 and smaller
than \code{n}.}

\item{both_tails}{Boolean. Do you consider both tails when computing the
causal tail coefficient? If \code{both_tails = TRUE}, then you use the
\eqn{\Psi} formulation. If \code{both_tails = FALSE},
then you use the \eqn{\Gamma} formulation.
Set by default to \code{TRUE}.}
}
\value{
Numeric vector. The causal order estimated from the data.
}
\description{
Runs extremal ancestral search (EASE) algorithm on the given
dataset \code{dat}.
}
