% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{fast_perm_search}
\alias{fast_perm_search}
\title{Fast permutation search}
\usage{
fast_perm_search(delta, mode = "sum", silent = FALSE)
}
\arguments{
\item{delta}{Numeric matrix --- between -1 and 1. The \code{delta}
matrix is defined as \code{gamma - t(gamma)}, where \code{gamma}
is the gamma coefficient matrix.}

\item{mode}{String. Is one of:
\itemize{
\item \code{"sum"} (default). At each step of the algorithm, choose
the node \eqn{i}, if the \emph{sum} of the entries of the \eqn{i}-th
row in \code{delta} is maximum across all the remaining nodes.
\item \code{"maxmin"}. At each step of the algorithm, choose the node
\eqn{i}, if the the \eqn{i}-th row of \code{delta} is the one which has
the maximum minimum (\emph{maxmin}) entry across all the remaining nodes.
}}

\item{silent}{Logical. Should the function communicate during the
execution?}
}
\value{
List. A list made of:
\itemize{
\item \code{order} Numeric vector. The causal order estimated
from \code{delta},
\item \code{score} Numeric. The maximized score associated to \code{order}.
It is defined as the sum of the upper triangular entries of \code{delta},
after its rows and columns are sorted according to \code{order}.
}
}
\description{
Runs fast permutation search given the matrix \code{delta}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
