% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_all_paths}
\alias{get_all_paths}
\title{Get all the directed paths of a DAG}
\usage{
get_all_paths(adj_mat)
}
\arguments{
\item{adj_mat}{Square numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Square numeric matrix. The path matrix associated to \code{adj_mat}.
}
\description{
Given the adjacency matrix \code{adj_mat}, produces a matrix
where the entry \eqn{(i, j)} measures the total weight of the paths
from node \eqn{i} to node \eqn{j}.
\strong{Note}: if \eqn{i = j}, the entry \eqn{(i, j) = 1}.
}
