% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_children}
\alias{get_children}
\title{Get the children of a DAG}
\usage{
get_children(g)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Numeric matrix. The children matrix of
the DAG \code{g}.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1} if \eqn{i} is a child
of \eqn{j} in the DAG \code{g}, and \eqn{0} otherwise.
}
