% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_descendants}
\alias{get_descendants}
\title{Get the descendants of a DAG}
\usage{
get_descendants(dag)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}
}
\value{
Square binary matrix. The descendant matrix of
the DAG \code{dag}.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1}
if \eqn{i} is a descendant of \eqn{j} in the DAG \code{dag},
and \eqn{0} otherwise.
}
