% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{greedy_perm_search}
\alias{greedy_perm_search}
\title{Greedy permutation search}
\usage{
greedy_perm_search(delta, silent = FALSE)
}
\arguments{
\item{delta}{Numeric matrix --- entries between -1 and 1. The \code{delta}
matrix is defined as \code{gamma - t(gamma)}, where \code{gamma}
is the gamma coefficient matrix. The number of rows (and columns)
of \code{delta} must be greater than 3.}

\item{silent}{Logical. Should the function communicate during the
execution?}
}
\value{
List. A list made of:
\itemize{
\item \code{order} Numeric vector. The causal order estimated
from \code{delta},
\item \code{score} Numeric. The maximized score associated to \code{order}.
It is defined as the sum of the upper triangular entries of \code{delta},
after its rows and columns are sorted according to \code{order}.
}
}
\description{
Runs greedy permutation search given the matrix \code{delta}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
