% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{inverse_mirror_uniform}
\alias{inverse_mirror_uniform}
\title{Inverse mirror uniform}
\usage{
inverse_mirror_uniform(prob, min, max)
}
\arguments{
\item{prob}{Numeric --- between 0 and 1. The probability associated
to the desired quantile.}

\item{min}{Numeric --- above 0. The lower bound of the positive half of
the support.}

\item{max}{Numeric --- above 0. The upper bound of the positive half of
the support. Note that \code{max} must be strictly greater than
\code{min}.}
}
\value{
Numeric --- between \code{-max} and \code{max}. The quantile
associated to the probability \code{prob}.
}
\description{
Produces the quantile of a mirrored uniform distribution
associated to the probability \code{prob}. The mirrored uniform
distribution has support [-max, -min] \eqn{\cup} [min, max].
}
