% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{lingam_search}
\alias{lingam_search}
\title{Lingam search}
\usage{
lingam_search(mat)
}
\value{
Numeric matrix (or \code{NA} in case of error).
The estimated adjacency matrix of a DAG.
}
\description{
Runs Lingam given a dataset \code{mat}.
}
