% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{minimax_search}
\alias{minimax_search}
\title{Minimax permutation search}
\usage{
minimax_search(gamma)
}
\arguments{
\item{gamma}{Numeric matrix --- between 0 and 1. The \code{gamma}
coefficient matrix.}
}
\value{
Numeric vector. The causal order estimated from \code{gamma}.
}
\description{
Runs minimax permutation search given the matrix \code{gamma}.
}
