% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{nonlinear_scm}
\alias{nonlinear_scm}
\title{Generate data from non-linear Structural Equation Model}
\usage{
nonlinear_scm(adj_mat, noise)
}
\arguments{
\item{adj_mat}{Square numeric matrix. The adjacency matrix of a DAG.}

\item{noise}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}
}
\value{
Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).
}
\description{
Generate data from non-linear Structural Causal Model as shown in the
paper "Causality in heavy-tailed models".
}
