% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{oracle_search}
\alias{oracle_search}
\title{Oracle permutation search}
\usage{
oracle_search(g)
}
\arguments{
\item{g}{Numeric matrix. The weighted adjacency matrix of a DAG.}
}
\value{
Numeric vector. The causal order estimated from \code{g}.
}
\description{
Runs oracle permutation search given the true DAG \code{g}.
}
