% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{order_lingam_search}
\alias{order_lingam_search}
\title{Order-Lingam search}
\usage{
order_lingam_search(dat, contrast_fun = c("logcosh", "exp"))
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{contrast_fun}{Character. The functional form of the contrast
function used in the Fast-ICA step. It is one of \code{"logcosh"}
(the default choice) and \code{"exp"}.
For further details see the paper from
Hyvarinen, A., \url{https://ieeexplore.ieee.org/abstract/document/761722/}.}
}
\value{
Numeric vector (or \code{NA} in case of error).
The causal order estimated from the data.
}
\description{
Runs order-Lingam given a dataset \code{dat}. Returns the causal order
estimated from the data using the LiNGAM-ICA algorithm.
}
\details{
This function is a wrapper around \code{estLiNGAM} from the package
\code{pcalg} (see \url{https://CRAN.R-project.org/package=pcalg }).
The function \code{estLiNGAM} is slightly modified
to allow for different contrast functions in the fast-ICA step of
LiNGAM. To modify \code{estLiNGAM}, we included in this package
several internal functions from\code{pcalg}.
All the credits go to the authors of the \code{pcalg} package:

Markus Kalisch, Alain Hauser , Martin Maechler, Diego Colombo,
Doris Entner, Patrik Hoyer, Antti Hyttinen, Jonas Peters,
Nicoletta Andri, Emilija Perkovic, Preetam Nandy, Philipp Ruetimann,
Daniel Stekhoven, Manuel Schuerch, Marco Eigenmann.
}
