% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{pc_rank_search}
\alias{pc_rank_search}
\title{PC (rank) search}
\usage{
pc_rank_search(dat, alpha)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{alpha}{Numeric --- between 0 and 1. The significance level for the
individual conditional independence tests.}
}
\value{
Square binary matrix (or \code{NA} in case of error).
The CPDAG estimated from the data.
}
\description{
Runs PC (rank) algorithm given a dataset \code{dat}.
}
\details{
The PC (rank) algorithm has been proposed by Harris, N., and
Drton, M.,
\url{http://jmlr.csail.mit.edu/papers/volume14/harris13a/harris13a.pdf}

The function \code{pcalg:pc()} is called with the following
arguments:
\itemize{
\item \code{suffStat = list(C = 2 * sin(cor(dat, method = "spearman")
* pi/6), n = nrow(dat))},
\item \code{indepTest = gaussCItest},
\item \code{u2pd = "retry"} --- this ensures that the produced CPDAG
is extendible to a DAG,
\item \code{skel.method = "stable"}.
}
}
