% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{pc_search}
\alias{pc_search}
\title{PC search}
\usage{
pc_search(mat)
}
\arguments{
\item{mat}{Numeric matrix. Matrix with \code{n} rows (observations)
and \code{p} columns (variables).}
}
\value{
List. A list made of:
\itemize{
\item \code{adj_mat} Numeric matrix (or \code{NA} in case of error).
The estimated adjacency matrix of a CPDAG,
\item \code{order} Numeric vector (or \code{NA} in case of error).
The causal order obtained by removing cycles from \code{adj_mat}.
}
}
\description{
Runs PC algorithm given a dataset \code{mat}.
}
\details{
The function \code{pcalg:pc()} is called with the following
arguments:
\itemize{
\item \code{suffStat = list(C = cor(mat), n = NROW(mat))},
\item \code{indepTest = gaussCItest},
\item \code{alpha = 0.05},
\item \code{u2pd = "retry"} --- this makes sure that the produced CPDAG
is extendible to a DAG,
\item \code{skel.method = "stable"}.
}
}
