% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{pick_elements}
\alias{pick_elements}
\title{Randomly pick elements
Select each element of \code{vec} with probability \code{prob} and
produce a vector with the selected elements.}
\usage{
pick_elements(vec, prob)
}
\arguments{
\item{vec}{Numeric vector. Vector containing the elements to select.}

\item{prob}{Numeric --- between 0 and 1. The probability of selecting
each element of \code{vec}.}
}
\value{
Numeric vector. A vector with the selected elements
}
\description{
Randomly pick elements
Select each element of \code{vec} with probability \code{prob} and
produce a vector with the selected elements.
}
