% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{psi_coefficient}
\alias{psi_coefficient}
\title{Compute \eqn{\Psi}-coefficient}
\usage{
psi_coefficient(adj_mat, i, j, tail_index)
}
\arguments{
\item{adj_mat}{Square numeric matrix. The adjacency matrix of a DAG.
If \code{both_tails = FALSE}, all the entries must be non-negative.}

\item{i, j}{Integer. The indices of the nodes to consider to compute
the theoretical gamma coefficient.}

\item{tail_index}{Numeric. The tail-index of the noise distribution.}
}
\value{
Numeric --- between 0 and 1. The theoretical psi coefficient
between node \code{i} and \code{j} in the DAG \code{adj_mat}.
}
\description{
Computes the theoretical \eqn{\Psi}-coefficient defined
in the paper "Causal discovery in heavy-tailed models".
}
\details{
For more information see the documentation of
\code{\link{psi_matrix}}.
}
