% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{psi_matrix}
\alias{psi_matrix}
\title{Compute \eqn{\Psi}-matrix}
\usage{
psi_matrix(adj_mat, alpha)
}
\arguments{
\item{adj_mat}{Square numeric matrix. The adjacency matrix of a DAG.
If \code{both_tails = FALSE}, all the entries must be non-negative.}

\item{alpha}{Numeric. The tail-index of the noise distribution.}
}
\value{
Numeric matrix. The entry \eqn{(i, j)} contains the
\eqn{\Psi} coefficient between node \eqn{i} and node \eqn{j}.
The values on the main diagonal are set to \code{NA}.
}
\description{
Computes the theoretical \eqn{\Psi}-matrix defined
in the paper "Causality in heavy-tailed models".
}
\details{
The causal tail coefficient is defined in
the paper "Causality in heavy-tailed models" and has two formulations.
This function considers the \eqn{\Psi}-formulation, and assumes that the
scales of the noise variables are equal to one.
}
