% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{random_b}
\alias{random_b}
\title{Sample random coefficients}
\usage{
random_b(g, lB = 0.1, uB = 0.9, two_intervals = FALSE)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}

\item{lB}{Numeric. The lower bound for the uniform distribution.}

\item{uB}{Numeric. The upper bound for the uniform distribution.
It must be stricly greater than \code{lB}.}

\item{two_intervals}{Logical. Should the coefficient be sampled
from two symmetric uniform distributions?
If \code{two_intervals == TRUE}, \code{lB} and \code{uB} must be
positive.}
}
\value{
Numeric matrix. The weighted adjacency matrix of the
DAG \code{g}.
}
\description{
Sample random coefficients from uniform distribution
for the given DAG \code{g}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
