% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{random_perm_search}
\alias{random_perm_search}
\title{Random permutation search}
\usage{
random_perm_search(delta)
}
\arguments{
\item{delta}{Numeric matrix --- between -1 and 1. The \code{delta}
matrix is defined as \code{gamma - t(gamma)}, where \code{gamma}
is the gamma coefficient matrix.}
}
\value{
List. A list made of:
\itemize{
\item \code{order} Numeric vector. The causal order estimated from \code{delta},
\item \code{score} Numeric. The score associated to \code{order}.
It is defined as the sum of the upper triangular entries of \code{delta},
after its rows and columns are sorted according to \code{order}.
}
}
\description{
Runs random permutation search given the matrix \code{delta}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
