% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{random_search}
\alias{random_search}
\title{Random search}
\usage{
random_search(dat)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}
}
\value{
Square binary matrix. A random DAG.
}
\description{
Produces a random DAG with random sparsity with as many variables
as the number of columns in the dataset \code{dat}.
}
