% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{sample_uniform}
\alias{sample_uniform}
\title{Sample from uniform family}
\usage{
sample_uniform(n, min, max, mirror = FALSE)
}
\arguments{
\item{n}{Integer. The number of elements to sample.}

\item{min}{Numeric. The lower bound of the distribution support.
If \code{mirror == TRUE}, this represents the lower bound of
the positive half of the support.}

\item{max}{Numeric. The upper bound of the distribution support.
If \code{mirror == TRUE}, this represents the upper bound of
the positive half of the support.
Note that \code{max} must be strictly greater than \code{min}.}

\item{mirror}{Logical. Should the elements be sampled from a
mirrored uniform distribution? If \code{mirror == TRUE},
\code{min} and \code{max} must be strictly greater than 0.}
}
\value{
Numeric vector. A vector with the sampled elements.
}
\description{
Sample n elements from a uniform distribution with lower and upper
bound equal to \code{min} and \code{max}, respectively.
If \code{mirror == TRUE}, the elements are sampled from a mirrored
uniform distribution, see \code{\link{inverse_mirror_uniform}}.
}
