% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate data from a random DAG}
\usage{
simulate_data(n, p, prob_connect, distr = c("student_t", "log_norm",
  "norm")[1], tail_index = 1.5, has_confounder = FALSE,
  is_nonlinear = FALSE, has_uniform_margins = FALSE)
}
\arguments{
\item{n}{Positive integer. The number of observations, must be larger
than 1.}

\item{p}{Positive integer. The number of variables, must be larger
than 1.}

\item{prob_connect}{Numeric --- between 0 and 1. The probability that an edge
\eqn{i {\rightarrow} j} is added to the DAG.}

\item{distr}{Character. The distribution of the noise. It is one of:
\itemize{
\item \code{student_t}, in this case the user has to specify the
\code{tail_index}, i.e., the degrees of freedom,
\item \code{gaussian},
\item \code{log_normal}.
}}

\item{tail_index}{Positive numeric. The tail index, i.e., degrees
of freedom, of the noise.}

\item{has_confounder}{Boolean. Are there confounders in the system?}

\item{is_nonlinear}{Boolean. Is the data generated non linear?}

\item{has_uniform_margins}{Boolean. Are the variables rescaled uniformly
between 0 and 1?}
}
\value{
List. The list is made of:
\itemize{
\item \code{dataset} --- Numeric matrix. Dataset of simulated data with
\code{n} rows and \code{p} columns (note that the hidden variables are not
included in this matrix).
\item \code{dag} --- Square binary matrix. The generated DAG, including
both the observed variables and the confounders,
if \code{has_confounder = TRUE}.
\item \code{pos_confounders} --- Integer vector. Represents the position
of confounders (rows and columns) in \code{dag}.
If \code{has_confounder = FALSE}, then \code{pos_confounders = integer(0)}.
}
}
\description{
Generate data from a random DAG.
}
