% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simulate_noise}
\alias{simulate_noise}
\title{Simulate noise observations}
\usage{
simulate_noise(
  n,
  p,
  distr = c("student_t", "gaussian", "log_normal"),
  tail_index
)
}
\arguments{
\item{n}{Positive integer. The number of observations, must be larger
than 1.}

\item{p}{Positive integer. The number of variables, must be larger
than 1.}

\item{distr}{Character. The distribution of the noise. It is one of:
\itemize{
\item \code{student_t}, in this case the user has to specify the
\code{tail_index}, i.e., the degrees of freedom,
\item \code{gaussian},
\item \code{log_normal}.
}}

\item{tail_index}{Positive numeric. The tail index, i.e., degrees
of freedom, of the noise.}
}
\value{
Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).
}
\description{
Sample \code{n} observations for \code{p} independent noise variables
from a certain distribution \code{distr}.
}
