% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{uniform_margin}
\alias{uniform_margin}
\title{Transform to uniform margins}
\usage{
uniform_margin(v)
}
\arguments{
\item{v}{Numeric vector.}
}
\value{
Numeric vector.
}
\description{
Transform the numeric vector \code{x} to uniform margin
between 0 and 1.
}
