% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_functions.R
\name{causal_metrics}
\alias{causal_metrics}
\title{Causal evaluation metrics}
\usage{
causal_metrics(g, est_g, order)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}

\item{est_g}{Numeric matrix. The (estimated) adjacency matrix of a
DAG or CPDAG.}
}
\description{
Evaluate the output of some causal inference model,
see \code{\link{causal_discovery}}, with different metrics.
}
\references{
List. The list is made of:
\itemize{
\item \code{str_int_dist} Numeric --- between 0 and 1 --- (or \code{NA} if
\code{est_g} is \code{NA}). The structural intervention distance between
the DAG \code{g} and the estimated adjacency matrix \code{est_g}. See also
\code{\link{compute_str_int_distance}}.
\item \code{ancestral_dist} Numeric --- between 0 and 1 --- (or \code{NA} if
\code{order} is \code{NA}). The ancestral distance between the DAG \code{g}
and the estimated order \code{order}.
See also \code{\link{compute_ancestral_distance}}.
}
}
