% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{compute_ancestral_distance}
\alias{compute_ancestral_distance}
\title{Compute ancestral distance}
\usage{
compute_ancestral_distance(g, caus_order)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}

\item{caus_order}{Numeric vector (or vector of \code{NA}s).
Represents a causal order.}
}
\value{
Numeric (or \code{NA}, if \code{caus_order} is a vector
of \code{NA}s). The ancestral distance between the DAG \code{g} and
the causal order \code{caus_order}, between 0 and 1.
}
\description{
Computes the ancestral distance between the DAG \code{g} and
the causal order \code{caus_order}. The ancestral distance is defined
as the number of pairs \eqn{(i, j)} where \eqn{i} is an ancestor
of \eqn{j} and \eqn{i} is placed after \eqn{j} in the causal order
\code{caus_order}.
}
