% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_extreme_functions.R
\name{compute_gamma_matrix}
\alias{compute_gamma_matrix}
\title{Causal gamma matrix}
\usage{
compute_gamma_matrix(dat, k = floor(2 * n^0.4), both_tails = TRUE)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{k}{Integer. The number of extreme observations used to
compute the gamma coefficient. Set by default to \code{k = floor(2 * n^0.4)}.}

\item{both_tails}{Boolean. Do you consider both tails when computing gamma?
Set by default to \code{TRUE}.}
}
\value{
Numeric matrix. The entry \eqn{(i, j)} contains the
gamma coefficient between the \eqn{i}-th and the \eqn{j}-th
column of \code{dat}. If \eqn{i = j}, the value of the entry \eqn{(i, j)}
is set to \code{NA}.
}
\description{
Computes the gamma coefficient matrix given a dataset \code{dat},
and a threshold \code{k}.
}
