% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_all_paths}
\alias{get_all_paths}
\title{Get all the directed paths of a DAG}
\usage{
get_all_paths(g, type = c("count", "weighted")[1])
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}

\item{type}{String. Is one of:
\itemize{
\item \code{"count"} (default),
\item \code{"weighted"}.
}}
}
\value{
Numeric matrix. The path matrix of the DAG \code{g}.
}
\description{
Given the DAG \code{g}, produces a matrix where the entry \eqn{(i, j)}
\itemize{
\item counts the number of directed paths from node \eqn{i} to node \eqn{j},
if \code{type == "count"},
\item measures the total weight of the paths from node \eqn{i}
to node \eqn{j}, if \code{type == "weighted"}.
}
\strong{Note}: if \eqn{i = j}, the entry \eqn{(i, j) = 1}.
}
