% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Get the ancestors of a DAG}
\usage{
get_ancestors(g)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Numeric matrix. The ancestral matrix of the DAG \code{g}.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1}
if \eqn{i} is an ancestor of \eqn{j} in the DAG \code{g},
and \eqn{0} otherwise.
}
