% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_children}
\alias{get_children}
\title{Get the children of a DAG}
\usage{
get_children(dag)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}
}
\value{
Square binary matrix. The children matrix of
the DAG \code{dag}.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1} if \eqn{i} is a child
of \eqn{j} in the DAG \code{dag}, and \eqn{0} otherwise.
}
