% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_descendants}
\alias{get_descendants}
\title{Get the descendants of a DAG}
\usage{
get_descendants(g)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Numeric matrix. The descendant matrix of
the DAG \code{g}.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1}
if \eqn{i} is a descendant of \eqn{j} in the DAG \code{g},
and \eqn{0} otherwise.
}
