% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_theory_functions.R
\name{get_parents}
\alias{get_parents}
\title{Get the parents of a DAG}
\usage{
get_parents(dag)
}
\arguments{
\item{dag}{Square binary matrix. A matrix representing a DAG.}
}
\value{
Square binary matrix. The parental matrix of
the DAG \code{dag}, i.e., the DAG itself.
}
\description{
Produces a matrix where the entry \eqn{(i, j) = 1} if \eqn{i} is a parent
of \eqn{j} in the DAG \code{dag}, and \eqn{0} otherwise.
}
