% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{lingam_search}
\alias{lingam_search}
\title{Lingam search}
\usage{
lingam_search(mat)
}
\arguments{
\item{mat}{Numeric matrix. Matrix with \code{n} rows (observations)
and \code{p} columns (variables).}
}
\value{
List. A list made of:
\itemize{
\item \code{adj_mat} Numeric matrix (or \code{NA} in case of error).
The estimated adjacency matrix of a DAG,
\item \code{order} Numeric vector (or \code{NA} in case of error).
The causal order obtained from \code{adj_mat}.
}
}
\description{
Runs Lingam given a dataset \code{mat}.
}
