% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{lingam_search}
\alias{lingam_search}
\title{Lingam search}
\usage{
lingam_search(dat, contrast_fun = c("logcosh", "exp")[1])
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{contrast_fun}{Character. The functional form of the contrast
function used in the Fast-ICA step. It is one of \code{"logcosh"}
(the default choice) and \code{"exp"}.
For further details see the paper from
Hyvarinen, A., \url{https://ieeexplore.ieee.org/abstract/document/761722/}.}
}
\value{
Square binary matrix (or \code{NA} in case of error).
The DAG estimated from the data.
}
\description{
Runs Lingam given a dataset \code{dat}.
}
