% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{pc_search}
\alias{pc_search}
\title{PC search}
\usage{
pc_search(dat, alpha = 0.05)
}
\arguments{
\item{dat}{Numeric matrix. Dataset matrix with \code{n}
rows (observations) and \code{p} columns (variables).}

\item{alpha}{Numeric --- between 0 and 1. The significance level for the
individual conditional independence tests. By default it is set to 0.05.}
}
\value{
Square binary matrix (or \code{NA} in case of error).
The CPDAG estimated from the data.
}
\description{
Runs PC algorithm given a dataset \code{dat}.
}
\details{
The function \code{pcalg:pc()} is called with the following
arguments:
\itemize{
\item \code{suffStat = list(C = cor(dat), n = NROW(dat))},
\item \code{indepTest = gaussCItest},
\item \code{u2pd = "retry"} --- this ensures that the produced CPDAG
is extendible to a DAG,
\item \code{skel.method = "stable"}.
}
}
