% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{random_coeff}
\alias{random_coeff}
\title{Sample random coefficients}
\usage{
random_coeff(g, lb = 0.1, ub = 0.9, two_intervals = FALSE)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}

\item{lb}{Numeric. The lower bound of the coefficient that can be
sampled.}

\item{ub}{Numeric. The lower bound of the coefficient that can be
sampled. It must be stricly greater than \code{lb}.}

\item{two_intervals}{Logical. Should the coefficient be sampled
from two symmetric uniform distributions?
If \code{two_intervals == TRUE}, \code{lb} and \code{ub} must be
positive.}
}
\value{
Numeric matrix. The weighted adjacency matrix of the
DAG \code{g}.
}
\description{
Sample random coefficients from uniform distribution
for the given DAG \code{g}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
