% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{random_dag}
\alias{random_dag}
\title{Simulate random DAG}
\usage{
random_dag(p, prob_connect, caus_order = sample(p, p, replace = FALSE))
}
\arguments{
\item{p}{Integer --- greater than 0. Number of nodes.}

\item{prob_connect}{Numeric --- between 0 and 1. The probability that an edge
\eqn{i {\rightarrow} j} is added to the DAG.}

\item{caus_order}{Numeric vector. The causal order of the DAG.
If the argument is not provided it is generated randomly.}
}
\value{
Square binary matrix. A matrix representing the random DAG.
}
\description{
Simulates a directed acyclic graph (DAG) and returns its adjacency matrix.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
\keyword{internal}
