% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_inference_helpers.R
\name{random_perm_search}
\alias{random_perm_search}
\title{Random permutation search}
\usage{
random_perm_search(g)
}
\arguments{
\item{g}{Numeric matrix. The adjacency matrix of a DAG.}
}
\value{
Numeric vector. The causal order estimated from \code{g}.
}
\description{
Runs random permutation search given the matrix \code{delta}.
Copyright (c) 2013 Jonas Peters \email{peters@math.ku.dk}.
All rights reserved.
}
